+++ lib/ansible/modules/cloud/amazon/ec2_win_password.py
@@ -63,6 +63,10 @@ extends_documentation_fragment:
 
 requirements:
     - cryptography
+
+notes:
+    - As of Ansible 2.4, this module requires the python cryptography module rather than the
+      older pycrypto module.
 '''
 
 EXAMPLES = '''
@@ -130,7 +134,7 @@ def main():
 
     instance_id = module.params.get('instance_id')
     key_file = module.params.get('key_file')
-    key_passphrase = module.params.get('key_passphrase')
+    b_key_passphrase = to_bytes(module.params.get('key_passphrase'), errors='surrogate_or_strict')
     wait = module.params.get('wait')
     wait_timeout = int(module.params.get('wait_timeout'))
 
@@ -155,13 +159,13 @@ def main():
         module.fail_json(msg = "wait for password timeout after %d seconds" % wait_timeout)
 
     try:
-        f = open(key_file, 'r')
+        f = open(key_file, 'rb')
     except IOError as e:
         module.fail_json(msg = "I/O error (%d) opening key file: %s" % (e.errno, e.strerror))
     else:
         try:
             with f:
-                key = load_pem_private_key(f.read(), key_passphrase, BACKEND)
+                key = load_pem_private_key(f.read(), b_key_passphrase, BACKEND)
         except (ValueError, TypeError) as e:
             module.fail_json(msg = "unable to parse key file")